- Tree model: moved most of NodeId's implementation to cpp file. - base: added new to_string() function (taking a template parameter for maximum flexibility) to convert simple values to a string quickly. - Replaced all "%z.." occurances by stream formatting (or to_string calls), as this format parameter is not supported by VC. - Fixed some warnings. - Win: TUT: grt classes are now set to auto register at project level. This avoids many warnings about no registration. - TUT: these changes fix at least be_inspector_value, if not others.
[TESTS]
- Running be_inspector_value locally.

+++ library/grt/unit-tests/diff/diff_tree_test.cpp
@@ -85,7 +85,7 @@ TEST_FUNCTION(1)
   for (std::size_t c = diff_tree->count(), i = 0; i < c; i++)
   {
     bec::NodeId schema((int)i);
-    for (int j = 0; j  < diff_tree->count_children(schema); j++)
+    for (size_t j = 0; j  < diff_tree->count_children(schema); j++)
     {
       bec::NodeId object(diff_tree->get_child(schema, j));
       std::string name;
